WIP: Test framework for gesture recognition
DO NOT COMMIT - not intended for trunk

This is a framework for gesture recognition using accelerometers
on a Peppy. It provides a task that is regularly calling a
gesture recognition function with accelerometer data input in
common/motion_sense.c.

Anyone who wants to implement a gesture recognition algorithm
should modify places labeled 'GESTURE TODO'.

There are two ectool commands that are useful:
- gesturerecord prints out accelerometer data which can be recorded
for offline analyzation.
- gestureconfig allows loading configuration information into EC
RAM for us in gesture recognition.

Additionally, the lid open and lid close host events are commented
out to allow for running in S0 with the lid closed.

BUG=none
BRANCH=none
TEST=Tested ectool commands on Peppy instrumented with
accelerometers.

Change-Id: Iad63fae5e7f9dac45bd9d74b3550c5130b3f080d
Signed-off-by: Alec Berg <alecaberg@chromium.org>