Update patch set 2
Patch Set 2: Code-Review-2

(1 comment)

Thanks for making the changes.

This change looks okay accept for the second copyright header in battery.c. We're unable to accept this as a contribution to the Chromium OS project with that second header included. If necessary, please escalate this at Quanta so the project leads at each company can discuss this issue. Thanks!

Patch-set: 2
Label: Code-Review=-2