lm4: Remove 500k clock delay in clock_init()
We copied that delay because it seemed to be necessary on early LM4
chips to avoid glitching the UART.  But on current boards (e.g. rambi)
this does not seem to be necessary, and delays boot by 31ms.  So,
remove the delay.

BUG=chrome-os-partner:23794
BRANCH=rambi
TEST=boot system; see little to no glitching on EC uart, and system boots ok
     hibernate 1; see little to no glitching on EC uart, and system boots ok

Orig-Change-Id: I9d4b5927da4282e47e1b09be838104c64f25268c
Signed-off-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/185232
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
(cherry picked from commit 36d4ecb153518623945f0a8e89e4413938d38869)

Change-Id: Ica9447e527da9d2098c17890ee63deacb362c166