Pit: Turn FET5 off when system is in suspend mode
This patch is to turn FET5 off when system is in the suspend mode. we
should turn camera power(FET5 in pmu) during the S2R because this is the
check list of the reliability team from this year.

BUG=chrome-os-partner:25603
TEST=Running the camera application on pi and pit and then excute
the suspend_stress_test tool. Then, checking the camera application is
working or not.
(I've tested this over 1000 times. this is good to us.)
BRANCH=pit

Change-Id: I79100c059ee291cbd0963055f2aeb3e96c304f39
Signed-off-by: Jaehoon kim <jh228.kim@samsung.com>