Optimize memcpy
This speeds up memcpy by copying a word at a time if source and destination are
aligned in mod 4. That is, if n and m are a positive integer:

  4n -> 4m: aligned, 4x speed.
  4n -> 4m+1: misaligned.
  4n+1 -> 4m+1: aligned in mod 4, 4x speed.

Memcmp aligned locations:

  > memcmp 0x200055a0 0x20000 0x1000
  memcmp(200055a0,20000,1000) = 0x10

Memcpy for 0x1000 bytes:

  > memcpy 0x200055a0 0x20000 0x1000
  [52.781911] memcpy(200055a0,20000,1000) = [52.782693] 0x200055a0

Verify copy is successful:

  > memcmp 0x200055a0 0x20000 0x1000
  memcmp(200055a0,20000,1000) = 0x0

Memcpy from a misaligned source:

  > memcpy 0x200055a0 0x20001 0x1000
  [66.727592] memcpy(200055a0,20001,1000) = [66.730733] 0x200055a0

Note this is about 4 times as slow as aligned copy (.003141 v.s. .000782).
Verify the result:

  > memcmp 0x200055a0 0x20001 0x1000
  memcmp(200055a0,20001,1000) = 0x0

Memcpy locations aligned in mod 4:

  > memcpy 0x200055a1 0x20001 0x1000
  [87.731560] memcpy(200055a1,20001,1000) = [87.732344] 0x200055a1

This is as fast as aligned copy (.000784). Verify the result:

  > memcmp 0x200055a1 0x20001 0x1000
  memcmp(200055a1,20001,1000) = 0x0

TEST=Performed on Peppy as described above. make buildall -j passes.
BUG=chrome-os-partner:23720
BRANCH=none
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: Ic12260451c5efd0896d6353017cd45d29cb672db
Tested-by: Daisuke Nojiri <dnojiri@google.com>