Optimize memcpy
This speeds up memcpy by copying a word at a time if source and destination are
aligned in mod 4. That is, if n and m are a positive integer:

  4n -> 4m: aligned, 4x speed.
  4n -> 4m+1: misaligned.
  4n+1 -> 4m+1: aligned in mod 4, 4x speed.

Note misaligned case is also optimized. Unit test runs in 298 us on Peppy while
it used to take 475.

Ran the unit test on Peppy:

  > runtest
  ...
  Running test_memcpy... (speed gain: 298 -> 93 us) OK
  ...

Ran make buildall -j:

  ...
  Running test_memcpy... (speed gain: 59 -> 8 us) OK
  ...

TEST=Described above.
BUG=chrome-os-partner:23720
BRANCH=none
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: Ic12260451c5efd0896d6353017cd45d29cb672db
Tested-by: Daisuke Nojiri <dnojiri@google.com>