Update patch set 9
Patch Set 9:

I rebased it to ToT and verified it works on Peppy. Is this patch still wanted? If so, I'll test it on an STM32 device and may ask Vic to test it for mec1322.

Patch-set: 9
Reviewer: Gerrit User 1106949 <1106949@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0