Update patch set 1
Patch Set 1: Code-Review+2

This is okay for now. Randall is adding the ability for the host to tell the EC whether to disable WiFi power in S3. That way the kernel driver can do the right thing depending on the module installed. See https://chromium-review.googlesource.com/#/c/187273/2

Patch-set: 1
Label: Code-Review=+2