Enable manual control of TOUCHSCREEN_RESET_L
The charger task was holding this either on or off in S3, no matter what we
wanted. We really only need to set it at S3->S0 or S3->S5, or when the
lid opens or closes. The rest of the time we should be able to turn it off
and on with gpioset, for testing purposes.

BUG=chrome-os-partner:26502
BRANCH=ToT
TEST=manual

Check the state with

  gpioget TOUCHSCREEN_RESET_L

Open and close the lid, suspend the AP with powerd_dbus_suspend, etc. The
touchscreen should be on when the lid is open and the AP is in either S3 or
S0, off when the lid is closed or the AP is off.

Then

  gpioset TOUCHSCREEN_RESET_L 1
  gpioget TOUCHSCREEN_RESET_L
  gpioset TOUCHSCREEN_RESET_L 0
  gpioget TOUCHSCREEN_RESET_L

The change should persist as long as nothing else changes.

Change-Id: If7b6f809b1b28ae2699d0fbc6c9b2305fc57cbff
Signed-off-by: Bill Richardson <wfrichar@chromium.org>