Update patch set 4
Patch Set 4: Code-Review+1

You can put this in ToT for now.

But note that "big is close to a release" will not be accepted as a reason for holding off on any other commits to ToT.  That is, by putting big in ToT you are subject to the whims of the EC team refactoring things.  We'll keep big nominally compiling (make buildall will pass), but as we don't have big boards we won't be able to test.

Patch-set: 4
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1