glimmer: clapper: add I2C unwedge ability to accel ports
This adds the ability for clapper and glimmer to automatically detect
and unwedge the I2C bus on the accelerometer port.

BUG=chrome-os-partner:26136
BRANCH=rambi
TEST=manual test on glimmer:

Used servo board to perform a cold boot until I noticed that on boot
I got an I2C unwedge message on the prompt. Then ran i2cscan to make
sure accelerometer bus was working. This shows that on boot we detected
a hang and fixed the porblem.

I've also tested this on the trunk using the i2c_wedge command which
is not yet in this branch.

Change-Id: If23ffbd8b41d1c559894a8ba1b15eb1525a91a13
Signed-off-by: Alec Berg <alecaberg@chromium.org>