Update patch set 1
Patch Set 1: Code-Review+2

Please add me as a reviewer with dparker@chromium.org instead of "daveparker+DONTUSE@gmail.com" as that second address is invalid.

It would be nice if you could add a more descriptive commit message. Something like:

Modify battery LED behavior

Set battery LED to WHITE when the battery is charged, the system is on AC power, and the system is off or in suspend.

Patch-set: 1