Update patch set 1
Patch Set 1: Code-Review+1

I would suggest perhaps doing this just for pit unless other platforms show the same problem. For snow we don't actually use an i2c passthrough so this problem should not occur. For spring I suspect the host comms is so slow that we wouldn't notice the EC going away for a while.

I'm wondering what is the root cause of this? If the EC fails to respond to a command in time, should we be more lenient on that timeout? Or is it just a CPU usage problem - i.e. the battery command is pegging the CPU and the other tasks never get to run (in which case this sounds like a more general problem with the task system).

I'd also like to suggest (for future work later in the year) another look at the idea of handling many of the commands in an interrupt context. Most commands take almost no time to execute and with a little bit of care we could improve EC responsiveness. I implemented this on snow and it seemed to work well.

One other point - the console code was extremely inefficient. I think Randall did improve it by implementing DMA for the UART - was that both for input and output? If this has not been done yet, it should reduce task switching significantly (again a future suggestion).

Patch-set: 1
Label: Code-Review=+1