Update patch set 1
Patch Set 1:

Simon:

I can do it just for pit if needed (and I'll do that if Louis says that's what he'd prefer), but I'm not sure I understand why Console needs to be a high priority on any systems.  Console is only there for debugging and Host Command should be more important.

See the bug for some more details, but we were seeing "SPI not ready" on the EC.  I'm not 100% sure the sequence of events that explains why making the console task a high priority fixes things, though.  I'm going to keep digging.

For other tasks, hopefully someone on the EC team can comment on them.

Patch-set: 1