Update patch set 1
Patch Set 1:

FYI: when we're running 'battery 1000' I think the problem is pretty clear.  It's spewing data over the console and eats up all of the priority, so of course there's no time to process host commands.

I don't know 100% that this is the same root cause as the kernel problems, but it doesn't seem impossible that our printing of "[XXXXX.XXXXXX HC 0x9e]" at 115200 at just the wrong time could cause problems.

> len("[XXXXX.XXXXXX HC 0x9e]\n") = 23
> (1000./11520) * 23 = 2.0 ms

...so assuming that the console is actually blocking then we could spend 2 ms there waiting for characters to print out.  I'll double-check all of that...

Patch-set: 1