Update patch set 1
Patch Set 1:

Hi Doug,

Agreed that the console should not be the highest priority if it causes the host interface to fail - suggest discussing that with Randall.

Console output with DMA would not consume CPU time. I did 'encourage' us to use DMA for UART, just not sure where it got to.

I think your theory is that anytime the EC is 'busy' (with battery 10000 or something else) then it does not respond in time to the AP. Sounds reasonable. But we know the EC can get busy when talking on I2C or when flashing, so we have to have a sufficient timeout that the AP can deal with that.

I think it is worth checking the recovery time between the AP de-asserting SPI CS and re-asserting it for the next transaction. This is EC_SPI_RECOVERY_TIME_NS in the kernel. The EC needs to see SPI CS go high before it resets RX ready for the next transaction (this could be changed, but it is how the EC currently works). I can't really see how the EC could miss this as it is an interrupt, as GPIOs are priority 1 and UART is priority 2. Worth digging into that if possible.

Ignore my comments on the DMA - it looks like Randall has it fully implemented.

Patch-set: 1