Update patch set 1
Patch Set 1: Code-Review+1

Adjusting the priority makes sense for "battery 1000" case, which generates tons of console messages, although the occasional failure is still suspicious to me.

However, for the issue 26000, I am not sure if this helps. But we could try.

Since this is the branch for Pit, it doesn't matter if we change other board setting (correct me if I was wrong).

Patch-set: 1
Label: Code-Review=+1