Update patch set 1
Patch Set 1:

Note that the LED on the device will indicate the battery is charged even though it may not be full. 

Instead of this CL we could add a formal FORCED_IDLE state to allow LED code to decide if it wants to indicate a different state or not.

Patch-set: 1