rambi: glimmer: Disable keyboard to wake from sleep when lid is open
Added ability to disable the keyboard to wake from S3 when the lid
is open past a certain angle. This has been added to glimmer by
defining USE_LID_ANGLE_TO_DISABLE_KEY_SCAN_IN_SLEEP.

Also modified the lid angle calculation to include a reliability
flag which can be used to tell when the hinge aligns too closely
with gravity and the lid angle value is unreliable.

BUG=none
BRANCH=rambi
TEST=Tested on a glimmer:

In S3, verified that when the lid is open past ~180 deg, the keyboard
does not wake the machine. Also verified that if you align hinge with
gravity, the keyboard enabled/disabled status remains the same (since
we can't actually trust the lid angle value).

Change-Id: If1a1592d259902d38941936961854b81b3a75b95
Signed-off-by: Alec Berg <alecaberg@chromium.org>