Update patch set 1
Patch Set 1:

I like the idea of adding a throttling flag. 

For the idea of having a host command to show more verbose status, perhaps could we rev the "chargedump" command to have a "pretty" option?

chargestate pretty (default)
chargestate raw    (same as chargedump)
chargedump -> chargestate raw (alias for backwards compatibility).

I'll file a separate bug for this. This is not essential for BayTrail so it could be included with the pending re-write of the charge state machine.

Patch-set: 1