TEST-ONLY: Make the battery NAK us
This adds a quick and dirty test command that shows the battery NAKing
us.  You can run "i2cxfer doug 1000 100" to see it.

The NAK seems to be related to the mode switch between amps and watts.
I'm not 100% sure why the tps65090 access in the middle is needed but
it may just make the timing work out right to show the battery
problem.

BRANCH=None
BUG=chrome-os-partner:26994
TEST=i2cxfer doug 1000 100

Change-Id: I86f21543bdc2abf372e0f7d5f0bb68f507b396d7
Signed-off-by: Doug Anderson <dianders@chromium.org>