Update patch set 2
Patch Set 2:

If the AP just suspends (S0->S3, S3->S0), then yes, we want the EC to remember and restore the limits.

But if the AP shuts down completely, the EC needs to NOT do that. Otherwise, the next time the AP boots, the EC will reimpose the previous limits and the newly booted kernel won't know that. If the kernel wants the limits set at boot, it needs to set them itself.

As it is now:

AP is in SO, sets DPTF limits.
AP shuts down (S0->S3->S5-G3). At S0->S3, the EC remembers the previous limit and sets the current limit to -1.

AP boots again. G3->S5->S3->S0. At S3->S0, the EC restores the previous limit. The new kernel has no idea this has happened.

You need a hook at S3->S5 for the EC to forget the previous limit by setting dptf_limit_ma_bak to -1.

Patch-set: 2