Update patch set 1
Patch Set 1: Verified+1

The wake_task() in the deffered_function scheduler is today's problem but other code may do something similar in the future. The ADC and I2C handlers already need this logic to deal with task events they don't care about so it's better to put this logic in one place.

Patch-set: 1
Label: Verified=+1