Update patch set 2
Patch Set 2:

Todd, Willis, and Max think it's better to turn this off by default. Landing this on the rambi branch first so we can test it with next week's dogfood firmware.

If folks complain, we can either revert or add a config option to enable this on a per-board basis.

Patch-set: 2