Add charge_state_v2 algorithm for use by Samus
This is a complete rewrite of the charge_state task used by x86 platforms.
Rather than having a bunch of state-specific functions, each with their own
error handling and special cases, this is organized like so:

  Forever:
  1. Read everything we can from the battery and charger.
  2. Figure out what we'd like to do (including error handling).
  3. Allow for customization to override that.
  4. Do it.

There's a lot to this CL (and more to come), so this is preliminary and
won't be committed exactly as-is.

Things I'm still working on are marked with "HEY:".
Things I need to file bugs for are marked with "TODO(HEY)".

None of the battery and charger error responses are handled yet (search for
"PROBLEM").

I also need to figure out why some tests are failing.

BUG=chrome-os-partner:20881
BRANCH=ToT
TEST=manual

  make buildall -j

Undefine CONFIG_BATTERY_PRESENT_CUSTOM, plug and unplug the battery to be
sure the trickle charging logic is correct when it can't tell if the battery
is present.

Change-Id: I3935cd3b87f322eb52178f8a675a886c16b75d58
Signed-off-by: Bill Richardson <wfrichar@chromium.org>