Clapper: Disable keyboard in suspend if lid is open beyond ~180 deg
Added ability to disable the keyboard to wake from suspend when the lid
is outside a certain angle range.

BUG=none
BRANCH=None
TEST=Tested on a clapper:

In S3, verified that when the lid is open past ~180 deg, the keyboard
does not wake the machine. Also verified that if you align hinge with
gravity, the keyboard enabled/disabled status remains the same (since
we can't actually trust the lid angle value).

Change-Id: I349e0fbc8def41f0280e2fb40af24fab4d465e3c
Signed-off-by: Dave Parker <dparker@chromium.org>