Update patch set 3
Patch Set 3:

Devin, could you check with the battery vendor about the following thing?

When the cut-off command is sent to battery, does the battery still response to i2c  (if the AC is still on)?

I am asking because your battery is different to other batteries we used before. When cut-off is sent, the battery will turn off all outputs, but still response to EC. So that when the AC is plugged in again, the EC can still talk to battery.

Patch-set: 3