Update patch set 3
Patch Set 3:

Hi Devin,

Tickle charging (pre-charge) is handled in state_common() of common/charge_state_v1.c  (around line 259-287):

        /*
         * Try to revive ultra low voltage pack.  Charge
         * battery pack with minimum current and maximum
         * voltage for 30 seconds.
         */

In this case, the charger would try to wake up battery for 30 secs. This should work on your case. You could trace if the charger goes into here. If you prefer, you could send me a blaze and HP battery so that we can reproduce and debug in US next time.

Your code would return a battery info when battery is dead. This will mislead the EC that battery is still there.

By the way, just heads-up that we are switching to charger V2 soon. We re-factored the complicated charge state. V2 will be easier to debug.

Patch-set: 3