Update patch set 2
Patch Set 2:

For stack holders of Tegra boards, I am going to move the charger code to v2. V2 is a re-factor so that we have more clear design on charger and battery code. This will make our life easier (debug and new features -- like fast charge) in the future.

Bill is verifying his code on x86. So far so good. So, please feel free to file bug if you encounter any problem.

Patch-set: 2