Update patch set 1
Patch Set 1:

Just trying to understand how the charger task could get scheduled while the console is printing. Is this correct: An 'Enter' is pressed. If the previous bytes in the buffer form a valid command the command is run in the console task. cprintf's during the command are handled via uart tx interrupts. When the command finishes the console command sleeps until a new character is entered.

Patch-set: 1