Update patch set 1
Patch Set 1:

(2 comments)

Based on the changes you've made in extpower_kip.c I take it that you want to get rid of support for turbo mode. I think the code would be a lot cleaner if we just stripped out the turbo part entirely and just set the adapter input current limit. Can you try to remove the extra code and structs for turbo?

Patch-set: 1