Update patch set 1
Patch Set 1:

You can print a corrupt value if the battery task is in the middle of updating the params when the console task preempts it and reads them.

Seems like it'd be safer for the console task to use the shared data area.  See command_flash_write() for a sample implementation.

Patch-set: 1