Update patch set 2
Patch Set 2: Code-Review-1

(3 comments)

Thanks for the comments Vincent!

After discussions with Vincent, we decided instead of storing EXC_RETURN on the stack, which wastes 64 bits of stack space when all we really care about is one bit, that we'll use the least significant bit of the stack pointer itself to hold EXC_RETURN[4]. It's trickier and more complex, but saves stack space, and should be slightly faster.

Patch-set: 2
Label: Code-Review=-1