Update patch set 1
Patch Set 1:

Dear Lo

I have another question here, because the charger code had moved to v2. In "static int charge_request(int voltage, int current)" function, it oniy set voltage and current when it is  over 0. --> if (voltage > 0) r1 = charger_set_voltage(voltage); if (current > 0) r2 = charger_set_current(current);
When the pre-charge is timeout, the charger setting will keep pre-charger setting and never set to 0. In charger_state_v1.c. The condition is voltage and current != 1 and it will be OK. Is it the common design in v2?

Patch-set: 1