Update patch set 1
Patch Set 1:

Dear Lo

My test about v2 is not completed, but the CL is verified by others. When I use the v2, it will stop charge when we charge battery and can not charge the battery to be full. When I use v1, it is OK. I am still searching the root cause. But our PVT SMT will start at this week. I'm afraid to use the v2 to factory build. I want to move back to v1 and do more test about v2.

Patch-set: 1