Update patch set 1
Patch Set 1: Code-Review-1

We decide to reduce the stack usage instead of increasing stack size because the 'battery' is a console command and not used in a shipping machine. Increasing the stack size would occupy the memory. The CL is almost done. Sorry that I didn't include you into the issue.

https://chromium-review.googlesource.com/#/c/194875/

Patch-set: 1
Label: Code-Review=-1