Update patch set 1
Patch Set 1:

Hi Cloud,

Could you report in the issue tracker how to reproduce this? I am trying to reproduce on my side. Your input will be very helpful as well.

We are trying to push V2 because we don't want to maintain 2 charger code. Your inputs will be very helpful to us to improve the v2. Thanks.

Patch-set: 1