charger v2: supports charger watchdog
Some chargers support a timeout mechanism that it would stop charging
if no voltage/current setting comes from battery or EC. This is designed
for safety.

In charger v1, it always updates charger periodically. But in v2, we
only update charger when needed. Add a timer for the charger enabling
this function.

BUG=chrome-os-partner:28201,chrome-os-partner:28208
BRANCH=nyan
TEST=Plug AC and battery.
Wait for 10 mins and ensure the battery is charged normally.

Change-Id: I00d62f8afa2fe2627ea9259f11679ced02af897a
Signed-off-by: Louis Yung-Chieh Lo <yjlou@chromium.org>