charger v2: only configure charger when AC is on.
The CL fcf26a4 enabled periodically charge_request(). However, that
could fail and generates lots of error message in the EC console if
the AC is not on and charger refuses the request (even it is 0v/0mA).

BUG=none
BRANCH=tot,nyan
TEST=make runtests. Tested on big. Expect NO below annoying error message
  [1.353104 charge_request(0mV, 0mA)]
  [1.453170 charge_request(0mV, 0mA)]
  [1.553281 charge_request(0mV, 0mA)]
  [1.653317 charge_request(0mV, 0mA)]
in the follwing cases:
  AC on, battery attached, power on, then remove/plug in AC.
  AC on, battery attached, power on, then remove/plug in battery.
  AC on, battery removed, power on, then plug in and remove battery.
  AC off, battery attached, power on, then plug in and remove AC.
'chgstate' also shows good state. At final, charge for 10 mins.

Change-Id: Icc729c52246df1ecfb7f289b5078dbc122b20a74
Signed-off-by: Louis Yung-Chieh Lo <yjlou@chromium.org>