fruitpie: Change USB PD trasmit to use SPI master mode
Added conditional compile flag PD_TX_USES_SPI_MASTER which switches
fruitpie over to use SPI master for PD transmit. By default this
option is not defined because it requires a HW change to the board
(MOSI and MISO need to be swapped between the ST and the Silego).
The advantage of SPI master mode is at the end of the tranmission,
we don't have to send any dummy 0 bits.

When the option is set, the system clock is also reduced to 38.4MHz.

BUG=chrome-os-partner:28309
BRANCH=none
TEST=Tested by connecting two fruitpies together across CC1. One
fruitpie has been modified and is running PD TX in SPI master mode,
while the other is running PD TX in SPI slave mode. On EC console
ran pd charger on one board and pd dev on other board. Verified
that communication works with no errors. Ran for 10 minutes in each
configuration.

Change-Id: Idb7636f7f58fcaf4fba147272bc31cd44a9ce477
Signed-off-by: Alec Berg <alecaberg@chromium.org>